Updates to .NET 8, updates nuget libraries, Fixes Sewage Cases by changing flow type from int to float

+++ sources/SloCovidServer/SloCovidServer/Mappers/SewageWeeklyCasesMapper.cs
@@ -38,7 +38,7 @@ public class SewageWeeklyCasesMapper : Mapper
             var item = new SewageWeeklyCases(
                 date.Year, date.Month, date.Day,
                 fields[stationIndex],
-                GetInt(fields[flowIndex]),
+                GetFloat(fields[flowIndex]),
                 n3,
                 GetInt(fields[codIndex]),
                 cases,
