+++ sources/SloCovidServer/SloCovidServer/Models/SewageWeeklyCases.cs
@@ -1,6 +1,6 @@
 ﻿namespace SloCovidServer.Models;
 
-public record SewageWeeklyCases(int Year, int Month, int Day, string Station, int? Flow, SewageN3 N3,
+public record SewageWeeklyCases(int Year, int Month, int Day, string Station, float? Flow, SewageN3 N3,
     int? Cod, SewageCase Cases, float? Lat, float? Lon, string Region, int? Population, float? CoverageRatio) : IModelDate;
 
 public record SewageN3(float? Raw, float? Norm);
